DDF-4133-credential-service     Unit tests in solr-factory are passing
