+++ platform/solr/solr-factory/src/main/java/org/codice/solr/factory/impl/ConfigurationStore.java
@@ -79,7 +79,7 @@ public class ConfigurationStore {
 
   public String getDataDirectoryPath() {
 
-    return SolrFactorySettings.getSolrDataDir();
+    return PublicSolrSettings.getSolrDataDir();
     // TODO: Remove this method
     // return dataDirectoryPath;
   }
